Research Paper (docx)
I wrote a research paper in my econometrics class with co-author Yunshu Liu that aims to identify and analyze the primary factors influencing the trade balance among G-20 countries. We utilized a dataset from the World Bank, and focused on variables like trade balance, exchange rates, and infrastructure quality. We used a model linking trade balance to these factors, finding that higher oil prices negatively impact trade, and a stronger exchange rate discourages trade. We also found infrastructure's positive effect on trade to be statistically insignificant, which contrasted with the paper we reviewed.